backend/requirements.txt
boto3==1.34.131
botocore==1.34.131
flask==3.0.3
flask-cors==4.0.1
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.3.0
requests==2.31.0
python-dotenv==1.0.0